requirements.txt
asgiref==3.4.1
certifi==2021.10.8
click==8.0.2
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.6
django-heroku==0.3.1
dlib 
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.1.0
numpy 
Pillow==8.1.1
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
waitress==2.0.0
whitenoise==5.2.0
wincertstore==0.2
